requirements.txt
# requirements.txt

albumentations==1.4.3
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
easydict==1.13
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.3
fsspec==2024.3.1
gdown==5.1.0
h11==0.14.0
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-multipart==0.0.9
pyvirtualcam==0.11.1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.12
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wget==3.2
